+++ src/Abstracts/Validator.php
@@ -20,7 +20,9 @@ use Estasi\Validator\{
     Interfaces\Validator as ValidatorInterface,
     Traits\Validation
 };
+use OutOfBoundsException;
 
+use function compact;
 use function get_class;
 use function is_iterable;
 use function is_null;
@@ -35,6 +37,11 @@ use function substr;
 /**
  * Class Validator
  *
+ * @property-read Translator|null $translator
+ * @property-read Map             $errorMessagesAliases
+ * @property-read string|null     $errorValueAlias
+ * @property-read int             $errorMessageLength
+ * @property-read bool            $errorValueObscured
  * @package Estasi\Validator\Abstracts
  */
 abstract class Validator implements ValidatorInterface
@@ -48,7 +55,7 @@ abstract class Validator implements ValidatorInterface
 
     // names of constructor parameters to create via the factory
     public const OPT_TRANSLATOR             = 'translator';
-    public const OPT_ERROR_MESSAGES_ALIASES = 'errorsMessagesAliases';
+    public const OPT_ERROR_MESSAGES_ALIASES = 'errorMessagesAliases';
     public const OPT_ERROR_VALUE_ALIAS      = 'errorValueAlias';
     public const OPT_ERROR_MESSAGE_LENGTH   = 'errorMessageLength';
     public const OPT_ERROR_VALUE_OBSCURED   = 'errorValueObscured';
@@ -66,13 +73,9 @@ abstract class Validator implements ValidatorInterface
     protected const MESSAGE_VAR_VALUE          = 'value';
     protected const MESSAGE_VAR_TYPES_EXPECTED = 'types';
 
-    private ?Translator  $translator;
-    private Map          $errorMessagesAliases;
-    private ?string      $errorValueAlias;
-    private int          $errorMessageLength;
-    private bool         $errorValueObscured;
-    private Map          $errorMessages;
-    private Map          $errorMessagesVars;
+    private Map $properties;
+    private Map $errorMessages;
+    private Map $errorMessagesVars;
 
     /**
      * Validator constructor.
@@ -82,29 +85,77 @@ abstract class Validator implements ValidatorInterface
      * @param string|null                                   $errorValueAlias
      * @param int                                           $errorMessageLength
      * @param bool                                          $errorValueObscured
+     * @param array                                         $options
      */
     public function __construct(
-        ?Translator $translator = self::WITHOUT_TRANSLATOR,
-        ?iterable $errorMessagesAliases = self::NO_ALIASES_FOR_ERROR_MESSAGES,
-        ?string $errorValueAlias = self::NO_ALIAS_VALUE_IN_ERROR_MESSAGE,
-        int $errorMessageLength = self::ERROR_MESSAGE_WITHOUT_LENGTH_LIMIT,
-        bool $errorValueObscured = self::DO_NOT_OBSCURE_VALUE
+        ?Translator $translator,
+        ?iterable $errorMessagesAliases,
+        ?string $errorValueAlias,
+        int $errorMessageLength,
+        bool $errorValueObscured,
+        array $options = []
     ) {
-        $this->translator           = $translator;
-        $this->errorMessagesAliases = new Map($errorMessagesAliases ?? []);
-        $this->errorValueAlias      = $errorValueAlias;
-        $this->errorMessageLength   = $errorMessageLength < self::ERROR_MESSAGE_WITHOUT_LENGTH_LIMIT
+        $errorMessagesAliases = new Map($errorMessagesAliases ?? []);
+        $errorMessageLength   = $errorMessageLength < self::ERROR_MESSAGE_WITHOUT_LENGTH_LIMIT
             ? self::ERROR_MESSAGE_WITHOUT_LENGTH_LIMIT
             : $errorMessageLength;
-        $this->errorValueObscured   = $errorValueObscured;
+        // setting basic properties
+        $this->properties = new Map(
+            compact('translator', 'errorMessagesAliases', 'errorValueAlias', 'errorMessageLength', 'errorValueObscured')
+        );
+        // setting properties of a child class (a specific validator)
+        $this->properties->putAll($options);
+
         $this->initErrorMessagesTemplates([self::E_INVALID_TYPE => 'The data type is not valid. Expected: "%types%"!']);
         $this->initErrorMessagesVars([]);
+    }
+
+    public function __get($name)
+    {
+        if ($this->properties->hasKey($name)) {
+            return $this->properties->get($name);
+        }
+
+        throw new OutOfBoundsException(
+            sprintf('The "%s" property of the class "%s" is undefined!', $name, static::class)
+        );
+    }
+
+    /**
+     * Returns a list with the values of parameters for the constructor of the Estasi\Validator\Abstracts\Validator
+     * class
+     *
+     * @param iterable|null $options    Basic validator options, such as the Translator, the length of the error
+     *                                  message, hiding the value being checked, defining your own error messages, and
+     *                                  so on.
+     * @param array         $properties properties of a child class (a specific validator)
+     *
+     * @return array<int, mixed>
+     */
+    protected function createProperties(?iterable $options, array $properties = []): array
+    {
+        $default = new Map(
+            [
+                self::OPT_TRANSLATOR             => self::WITHOUT_TRANSLATOR,
+                self::OPT_ERROR_MESSAGES_ALIASES => self::NO_ALIASES_FOR_ERROR_MESSAGES,
+                self::OPT_ERROR_VALUE_ALIAS      => self::NO_ALIAS_VALUE_IN_ERROR_MESSAGE,
+                self::OPT_ERROR_MESSAGE_LENGTH   => self::ERROR_MESSAGE_WITHOUT_LENGTH_LIMIT,
+                self::OPT_ERROR_VALUE_OBSCURED   => self::DO_NOT_OBSCURE_VALUE,
+                self::OPT_OPTIONS                => $properties,
+            ]
+        );
+
+        return $default->merge(new Map($options ?? []))
+                       ->intersect($default)
+                       ->values()
+                       ->toArray();
     }
 
     /**
      * @param iterable<string, mixed>|null $options
      *
      * @return array<int, mixed>
+     * @deprecated
      */
     final protected function getValidOptionsForParent(iterable $options = null): array
     {